+++ docs/docs/en/api/fastagency/core/io/mesop/components/ui_common/header.md
@@ -8,4 +8,4 @@ search:
   boost: 0.5
 ---
 
-::: fastagency.core.io.mesop.Mesop
+::: fastagency.core.io.mesop.components.ui_common.header
