+++ docs/docs/en/api/fastagency/core/io/mesop/main/get_workflows.md
@@ -8,4 +8,4 @@ search:
   boost: 0.5
 ---
 
-::: fastagency.core.io.mesop.base.Mesop
+::: fastagency.core.io.mesop.main.get_workflows
