+++ docs/docs/en/api/fastagency/core/io/mesop/send_prompt/send_prompt_to_autogen.md
@@ -1,0 +1,11 @@
+---
+# 0.5 - API
+# 2 - Release
+# 3 - Contributing
+# 5 - Template Page
+# 10 - Default
+search:
+  boost: 0.5
+---
+
+::: fastagency.core.io.mesop.send_prompt.send_prompt_to_autogen
