+++ fastagency/core/io/mesop/data_model.py
@@ -1,17 +1,19 @@
 from dataclasses import dataclass, field
-from typing import  Optional
+from typing import Optional
+
 import mesop as me
 
 
 @dataclass
 class Conversation:
-    #messages: list[ConversationMessage] = field(default_factory=list)
+    # messages: list[ConversationMessage] = field(default_factory=list)
     messages: list[str] = field(default_factory=list)
+
 
 @me.stateclass
 class State:
-    conversationCompleted: bool = False
-    waitingForFeedback: bool = False
+    conversation_completed: bool = False
+    waiting_for_feedback: bool = False
     prompt: str = ""
     feedback: str = ""
     conversation: Conversation
