+++ fastagency/core/io/mesop/send_prompt.py
@@ -2,14 +2,16 @@ from collections.abc import Iterable
 
 import mesop as me
 
-from examples.mesop_poc.data_model import State
-from examples.mesop_poc.workflows import wf
-from fastagency.core.mesop.base import MesopIO, MesopMessage, run_workflow
+from ...base import Workflows
+from .base import MesopIO, MesopMessage, run_workflow
+from .data_model import State
 
 
-def send_prompt_to_autogen(prompt: str) -> Iterable[MesopMessage]:
+def send_prompt_to_autogen(
+    prompt: str, wf: Workflows, name: str
+) -> Iterable[MesopMessage]:
     # mesop_io = run_workflow(wf, "simple_learning", prompt)
-    mesop_io = run_workflow(wf, "exam_practice", prompt)
+    mesop_io = run_workflow(wf, name=name, initial_message=prompt)
     state = me.state(State)
     state.fastagency = mesop_io.id
     return mesop_io.get_message_stream()
